// app/types.ts
 export type ProductType = {
   id: string;
   product_id: string;
   product: string;
   tags: string;
   DEHP_equivalents_ng_g: string | number;
   DEHP_ng_g: string | number;
   DBP_ng_g: string | number;
   BBP_ng_g: string | number;
   DINP_ng_g: string | number;
   DIDP_ng_g: string | number;
   blinded_photo: string;
   collected_at: string;
   expiration_date: string;
   analysis_method_phthalates: string;
   DEHP_ng_serving: string | number;
   DBP_ng_serving: string | number;
   BBP_ng_serving: string | number;
   DINP_ng_serving: string | number;
   DIDP_ng_serving: string | number;
   DEHP_percent_tdi_14_kg_efsa: string | number;
   DBP_percent_tdi_14_kg_efsa: string | number;
   BBP_percent_tdi_14_kg_efsa: string | number;
   DINP_percent_tdi_14_kg_efsa: string | number;
   DIDP_percent_tdi_14_kg_efsa: string | number;
   DEHP_percent_tdi_14_kg_epa: string | number;
   DBP_percent_tdi_14_kg_epa: string | number;
   BBP_percent_tdi_14_kg_epa: string | number;
   DINP_percent_tdi_14_kg_epa: string | number;
   DIDP_percent_tdi_14_kg_epa: string | number;
   DEHP_percent_tdi_70_kg_efsa: string | number;
   DBP_percent_tdi_70_kg_efsa: string | number;
   BBP_percent_tdi_70_kg_efsa: string | number;
   DINP_percent_tdi_70_kg_efsa: string | number;
   DIDP_percent_tdi_70_kg_efsa: string | number;
   DEHP_percent_tdi_70_kg_epa: string | number;
   DBP_percent_tdi_70_kg_epa: string | number;
   BBP_percent_tdi_70_kg_epa: string | number;
   DINP_percent_tdi_70_kg_epa: string | number;
   DIDP_percent_tdi_70_kg_epa: string | number;
   BPA_ng_g: string | number;
   BPS_ng_g: string | number;
   BPF_ng_g: string | number;
   BPA_ng_serving: string | number;
   BPS_ng_serving: string | number;
   BPF_ng_serving: string | number;
   BPA_percent_tdi_14_kg_efsa: string | number;
   BPS_percent_tdi_14_kg_efsa: string | number;
   BPF_percent_tdi_14_kg_efsa: string | number;
   BPA_percent_tdi_14_kg_epa: string | number;
   BPS_percent_tdi_14_kg_epa: string | number;
   BPF_percent_tdi_14_kg_epa: string | number;
   BPA_percent_tdi_70_kg_efsa: string | number;
   BPS_percent_tdi_70_kg_efsa: string | number;
   BPF_percent_tdi_70_kg_efsa: string | number;
   BPA_percent_tdi_70_kg_epa: string | number;
   BPS_percent_tdi_70_kg_epa: string | number;
   BPF_percent_tdi_70_kg_epa: string | number;
   DEHT_ng_g: string | number;
   DEHA_ng_g: string | number;
   DINCH_ng_g: string | number;
   DIDA_ng_g: string | number;
   DEHT_ng_serving: string | number;
   DEHA_ng_serving: string | number;
   DINCH_ng_serving: string | number;
   DIDA_ng_serving: string | number;
   DEHT_percent_tdi_14_kg_efsa: string | number;
   DEHA_percent_tdi_14_kg_efsa: string | number;
   DINCH_percent_tdi_14_kg_efsa: string | number;
   DIDA_percent_tdi_14_kg_efsa: string | number;
   DEHT_percent_tdi_14_kg_epa: string | number;
   DEHA_percent_tdi_14_kg_epa: string | number;
   DINCH_percent_tdi_14_kg_epa: string | number;
   DIDA_percent_tdi_14_kg_epa: string | number;
   DEHT_percent_tdi_70_kg_efsa: string | number;
   DEHA_percent_tdi_70_kg_efsa: string | number;
   DINCH_percent_tdi_70_kg_efsa: string | number;
   DIDA_percent_tdi_70_kg_efsa: string | number;
   DEHT_percent_tdi_70_kg_epa: string | number;
   DEHA_percent_tdi_70_kg_epa: string | number;
   DINCH_percent_tdi_70_kg_epa: string | number;
   DIDA_percent_tdi_70_kg_epa: string | number;
   serving_size_g: string | number;
 };